Smoother RSS feed generation on user side
ITP-8
RSS feed URLs are now handled in part by the URL mapping
object in the Grails project. This means they can be of
the form .../RSS/gameName/playerName rather than the form
.../RSS?gameName=foo&playerName=bar. Additionally, as seen
above, the new format uses strings to specify games and
players by rom filenames and player names, respectively.

The RSS indicator on the show pages is changed to an icon
at the bottom of the page rather than an inline text note.